Rename mod from origin to origin-domain
